misc refactor to controller and router
